refactored code and addded null functionality to final page for emails
